requirements.txt
Django==4.2
stripe>=5.0.0
python-dotenv==1.0.0